+++ packages/adapter-mongoose/lib/adapter-mongoose.js
@@ -8,19 +8,11 @@ const {
   mapKeyNames,
   identity,
   mergeWhereClause,
-} = require('@keystone-alpha/utils');
+} = require('@keystone/utils');
 
-const {
-  BaseKeystoneAdapter,
-  BaseListAdapter,
-  BaseFieldAdapter,
-} = require('@keystone-alpha/keystone');
-const {
-  queryParser,
-  pipelineBuilder,
-  mutationBuilder,
-} = require('@keystone-alpha/mongo-join-builder');
-const logger = require('@keystone-alpha/logger').logger('mongoose');
+const { BaseKeystoneAdapter, BaseListAdapter, BaseFieldAdapter } = require('@keystone/keystone');
+const { queryParser, pipelineBuilder, mutationBuilder } = require('@keystone/mongo-join-builder');
+const logger = require('@keystone/logger').logger('mongoose');
 
 const slugify = require('@sindresorhus/slugify');
 
@@ -81,7 +73,7 @@ class MongooseAdapter extends BaseKeystoneAdapter {
 
   getDefaultPrimaryKeyConfig() {
     // Required here due to circular refs
-    const { MongoId } = require('@keystone-alpha/fields-mongoid');
+    const { MongoId } = require('@keystone/fields-mongoid');
     return MongoId.primaryKeyDefaults[this.name].getConfig();
   }
 }