+++ packages/field-content/src/blocks/link.js
@@ -1,4 +1,4 @@
-import { importView } from '@keystone-alpha/build-field-types';
+import { importView } from '@keystone/build-field-types';
 import { Block } from '../Block';
 
 export default class LinkBlock extends Block {