+++ packages/fields/src/types/Decimal/Implementation.js
@@ -1,7 +1,7 @@
 import mongoose from 'mongoose';
 import { Implementation } from '../../Implementation';
-import { MongooseFieldAdapter } from '@keystone-alpha/adapter-mongoose';
-import { KnexFieldAdapter } from '@keystone-alpha/adapter-knex';
+import { MongooseFieldAdapter } from '@keystone/adapter-mongoose';
+import { KnexFieldAdapter } from '@keystone/adapter-knex';
 
 export class Decimal extends Implementation {
   constructor(path, { symbol }) {