+++ packages/test-utils/lib/test-utils.js
@@ -3,10 +3,10 @@ const supertest = require('supertest-light');
 const MongoDBMemoryServer = require('mongodb-memory-server').default;
 const pFinally = require('p-finally');
 const url = require('url');
-const { Keystone } = require('@keystone-alpha/keystone');
-const { GraphQLApp } = require('@keystone-alpha/app-graphql');
-const { KnexAdapter } = require('@keystone-alpha/adapter-knex');
-const { MongooseAdapter } = require('@keystone-alpha/adapter-mongoose');
+const { Keystone } = require('@keystone/keystone');
+const { GraphQLApp } = require('@keystone/app-graphql');
+const { KnexAdapter } = require('@keystone/adapter-knex');
+const { MongooseAdapter } = require('@keystone/adapter-mongoose');
 
 async function setupServer({
   name,