+++ packages/adapter-knex/lib/adapter-knex.js
@@ -117,7 +117,7 @@ class KnexAdapter extends BaseKeystoneAdapter {
   // This will completely drop the backing database. Use wisely.
   dropDatabase() {
     const tables = Object.values(this.listAdapters)
-      .map(listAdapter => `"${this.schemaName}"."${listAdapter.key}"`)
+      .map(listAdapter => `"${this.schemaName}"."${listAdapter.tableName}"`)
       .join(',');
     return this.knex.raw(`DROP TABLE IF EXISTS ${tables} CASCADE`);
   }
@@ -138,6 +138,7 @@ class KnexListAdapter extends BaseListAdapter {
     super(...arguments);
     this.getListAdapterByKey = parentAdapter.getListAdapterByKey.bind(parentAdapter);
     this.realKeys = [];
+    this.tableName = this.key;
   }
 
   prepareFieldAdapter(fieldAdapter) {
@@ -160,7 +161,7 @@ class KnexListAdapter extends BaseListAdapter {
 
   async createTable() {
     // Let the field adapter add what it needs to the table schema
-    await this._schema().createTable(this.key, table => {
+    await this._schema().createTable(this.tableName, table => {
       this.fieldAdapters.forEach(adapter => adapter.addToTableSchema(table));
     });
   }
@@ -169,7 +170,7 @@ class KnexListAdapter extends BaseListAdapter {
     const relationshipAdapters = this.fieldAdapters.filter(adapter => adapter.isRelationship);
 
     // Add foreign key constraints on this table
-    await this._schema().table(this.key, table => {
+    await this._schema().table(this.tableName, table => {
       relationshipAdapters
         .filter(adapter => !adapter.config.many)
         .forEach(adapter => adapter.createForeignKey(table, this.parentAdapter.schemaName));
@@ -221,7 +222,7 @@ class KnexListAdapter extends BaseListAdapter {
       table
         .foreign(leftFkPath)
         .references(leftPkFa.path) // 'id'
-        .inTable(`${dbAdapter.schemaName}.${leftListAdapter.key}`)
+        .inTable(`${dbAdapter.schemaName}.${leftListAdapter.tableName}`)
         .onDelete('CASCADE');
 
       rightPkFa.addToForeignTableSchema(table, {
@@ -233,7 +234,7 @@ class KnexListAdapter extends BaseListAdapter {
       table
         .foreign(rightFkPath)
         .references(rightPkFa.path) // 'id'
-        .inTable(`${dbAdapter.schemaName}.${rightListAdapter.key}`)
+        .inTable(`${dbAdapter.schemaName}.${rightListAdapter.tableName}`)
         .onDelete('CASCADE');
     });
   }
@@ -243,7 +244,7 @@ class KnexListAdapter extends BaseListAdapter {
     const realData = pick(data, this.realKeys);
     const item = (await this._query()
       .insert(realData)
-      .into(this.key)
+      .into(this.tableName)
       .returning('*'))[0];
 
     // For every many-field, update the many-table
@@ -289,7 +290,7 @@ class KnexListAdapter extends BaseListAdapter {
                     .del()
                 : adapter
                     ._query()
-                    .table(adapter.key)
+                    .table(adapter.tableName)
                     .where(a.path, id)
                     .update({ [a.path]: null })
             )
@@ -298,7 +299,7 @@ class KnexListAdapter extends BaseListAdapter {
     );
     // Delete the actual item
     return this._query()
-      .table(this.key)
+      .table(this.tableName)
       .where({ id })
       .del();
   }
@@ -328,7 +329,7 @@ class KnexListAdapter extends BaseListAdapter {
     // Update the real data
     const realData = pick(data, this.realKeys);
     const query = this._query()
-      .table(this.key)
+      .table(this.tableName)
       .where({ id });
     if (Object.keys(realData).length) {
       query.update(realData);
@@ -388,7 +389,7 @@ class KnexListAdapter extends BaseListAdapter {
   async _findById(id) {
     return (
       (await this._query()
-        .from(this.key)
+        .from(this.tableName)
         .where('id', id))[0] || null
     );
   }
@@ -439,7 +440,7 @@ class QueryBuilder {
     this._tableAliases = {};
     this._nextBaseTableAliasId = 0;
     const baseTableAlias = this._getNextBaseTableAlias();
-    this._query = listAdapter._query().from(`${listAdapter.key} as ${baseTableAlias}`);
+    this._query = listAdapter._query().from(`${listAdapter.tableName} as ${baseTableAlias}`);
     if (meta) {
       this._query.count();
     } else {
@@ -517,7 +518,7 @@ class QueryBuilder {
           if (!this._tableAliases[otherTableAlias]) {
             this._tableAliases[otherTableAlias] = true;
             query.leftOuterJoin(
-              `${otherList} as ${otherTableAlias}`,
+              `${otherListAdapter.tableName} as ${otherTableAlias}`,
               `${otherTableAlias}.id`,
               `${tableAlias}.${path}`
             );
@@ -572,7 +573,7 @@ class QueryBuilder {
             .select(`${subBaseTableAlias}.${thisID}`)
             .from(`${manyTableName} as ${subBaseTableAlias}`);
           subQuery.innerJoin(
-            `${otherListAdapter.key} as ${otherTableAlias}`,
+            `${otherListAdapter.tableName} as ${otherTableAlias}`,
             `${otherTableAlias}.id`,
             `${subBaseTableAlias}.${otherList}_id`
           );