+++ packages/oembed-adapters/CHANGELOG.md
@@ -1,3 +1,5 @@
+# @keystone/oembed-adapters
+
 # @keystone-alpha/oembed-adapters
 
 ## 1.1.2